OSIS-927 clean internship submodule
