+++ base/models/academic_year.py
@@ -23,9 +23,12 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+import datetime
+
 from django.db import models
 from django.utils import timezone
 
+from base import models as mdl_base
 from osis_common.models.serializable_model import SerializableModel, SerializableModelAdmin
 
 LEARNING_UNIT_CREATION_SPAN_YEARS = 6
@@ -119,3 +122,9 @@ def starting_academic_year():
 
 def compute_max_academic_year_adjournment():
     return starting_academic_year().year + LEARNING_UNIT_CREATION_SPAN_YEARS
+
+
+def get_last_academic_years(last_years=10):
+    today = datetime.date.today()
+    date_ten_years_before = today.replace(year=today.year - last_years)
+    return mdl_base.academic_year.find_academic_years().filter(start_date__gte=date_ten_years_before)
